# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "\nТобі потрібно реалізувати сервіс з АРІ, який дозволить: \n- дізнатись поточний курс біткоіну (BTC) у гривні (UAH)\n- підписати емейл на отримання інформації по зміні курсу\n- запит, який відправить всім підписаним користувачам актуальний курс. \n\nДодаткові вимоги:\n1. Сервіс має відповідати описаному ниже АРІ. <i>NB Закривати рішення аутентифікацією не потрібно</i>. \n2. Всі данні, для роботи додатку повинні зберігатися в файлах (підключати базу данних не потрібно). Тобто, потрібно реалізувати збереження та роботу з даними (наприклад, електронними адресами) через файлову систему.\n3. В репозиторії повинен бути Dockerfile, який дозволяє запустити систему в Docker. З матеріалом по Docker вам необхідно ознайомитись самостійно.\n4. Також ти можеш додавати коментарі чи опис логіки виконання роботи в README.md документі. Правильна логіка може стати перевагою при оцінюванні, якщо ти не повністю виконаєш завдання.\n\nОчікувані мови виконання завдання: PHP, Go, JavaScript (Node.js).  Виконувати завдання іншими мовами можна, проте, це не буде перевагою.\nВиконане завдання необхідно завантажити на GitHub (публічний репозиторій) та сабмітнути виконане завдання в гугл-форму.\n\nТи можеш користуватися усією доступною інформацією, але виконуй завдання самостійно. \nУспіхів! ",
    "version": "1.0.0",
    "title": "GSES2 BTC application"
  },
  "basePath": "/api",
  "tags": [
    {
      "name": "rate",
      "description": "Отримання поточного курсу BTC до UAH"
    },
    {
      "name": "subscription",
      "description": "Робота з підпискою"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/rate": {
      "get": {
        "tags": [
          "rate"
        ],
        "summary": "Отримати поточний курс BTC до UAH",
        "description": "Запит має повертати поточний курс BTC до UAH використовуючи будь-який third party сервіс з публічним АРІ",
        "operationId": "rate",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Повертається актуальний курс BTC до UAH",
            "schema": {
              "type": "integer"
            }
          },
          "400": {
            "description": "Помилка виконання запиту"
          }
        }
      }
    },
    "/subscribe": {
      "post": {
        "tags": [
          "subscription"
        ],
        "summary": "Підписати емейл на отримання поточного курсу",
        "description": "Запит має перевірити, чи немає данної електронної адреси в поточній базі даних (файловій) і, в разі її відсутності, записувати її. Пізніше, за допомогою іншого запиту ми будемо відправляти лист на ті електронні адреси, які будуть в цій базі. ",
        "operationId": "subscribe",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "Електронна адреса, яку потрібно підписати",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "E-mail додано"
          },
          "409": {
            "description": "Повертати, якщо e-mail вже є в базі даних (файловій)"
          },
          "400": {
            "description": "(Необов'язковий) Невдала валідація пошти"
          }
        }
      }
    },
    "/sendEmails": {
      "post": {
        "tags": [
          "subscription"
        ],
        "summary": "Відправити e-mail з поточним курсом на всі підписані електронні пошти.",
        "description": "Запит має отримувати актуальний курс BTC до UAH за допомогою third-party сервісу та відправляти його на всі електронні адреси, які були підписані раніше.  ",
        "operationId": "sendEmails",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "E-mailʼи відправлено"
          },
          "204": {
            "description": "(Необов'язковий) В базі відсутня пошта"
          }
        }
      }
    },
    "/unsubscribe": {
      "delete": {
        "tags": [
          "subscription"
        ],
        "summary": "Відписатися від отримання поточного курсу",
        "description": "Запит має видалити пошту зі списку для отримання курсу.",
        "operationId": "unsubscribe",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "Електронна адреса, яку потрібно відписати",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "E-mail видалено"
          },
          "409": {
            "description": "Пошта вже видалена з бази даних"
          },
          "400": {
            "description": "(Необов'язковий) Невдала валідація пошти"
          }
        }
      }
    },
    "/subscribers": {
      "get": {
        "tags": [
          "subscription"
        ],
        "summary": "Список підписаних на отримання поточного курсу",
        "description": "Запит має повернути масив з поштами, що є підписаними на отримання курсу.",
        "operationId": "subscribers",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Список електроних пошт"
          }
        }
      }
    },
    "/changeProviderByName": {
      "put": {
        "tags": [
          "rate"
        ],
        "summary": "Змінити постачальника даних курсу по імені провайдера.",
        "description": "Запит має змінити постачальника(third-party сервіс) курсу валют та повідомити про успіх.",
        "operationId": "changeProviderByName",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "formData",
            "description": "Ім'я провайдера (З великої літери)",
            "required": true,
            "type": "string",
            "enum": ["Binance", "Test", "[WrongToTest]"]
          }
        ],
        "responses": {
          "200": {
            "description": "Провайдер успішно змінено"
          },
          "404": {
            "description": "Провайдер не знайдено"
          }
        }
      }
    },
    "/changeProviderByKey": {
      "put": {
        "tags": [
          "rate"
        ],
        "summary": "Змінити постачальника даних курсу по коду провайдера.",
        "description": "Запит має змінити постачальника(third-party сервіс) курсу валют та повідомити про успіх.",
        "operationId": "changeProviderByKey",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "formData",
            "description": "Код провайдера",
            "required": true,
            "type": "string",
            "enum": ["binance-pk", "test-pk", "[WrongToTest]"]
          }
        ],
        "responses": {
          "200": {
            "description": "Провайдер успішно змінено"
          },
          "404": {
            "description": "Провайдер не знайдено"
          }
        }
      }
    }
  }
}